+++ recipes-support/irma6-swuimage/files/pre_post_inst.sh
@@ -7,7 +7,7 @@ fi
 TAG=$0
 
 log() {
-	logger -t $TAG $1
+	logger -t "$TAG" "$1"
 }
 
 exists() {
@@ -100,11 +100,56 @@ umount_keystore() {
 	umount ${KEYSTORE}
 }
 
+resize_lvm() {
+	# suppress lvm tool warnings regarding closing of all file descriptors
+	export LVM_SUPPRESS_FD_WARNINGS=1
+
+	# get new compressed/encrypted rootfs
+	if [ ! -e /tmp/sw-description ]; then
+		log "Could not find sw-description file during logical volume resize!"; exit 1;
+	fi
+	rootfs_file=$(grep ext4.gz /tmp/sw-description | grep -Eo "\".*\"" | sed 's/"//g')
+	if [ -z "$rootfs_file" ]; then
+		log "Could not find new rootfs during logical volume resize!"; exit 1;
+	fi
+	rootfs_file="/tmp/$rootfs_file"
+	if [ ! -e "$rootfs_file" ]; then
+		log "Could not find new rootfs during logical volume resize!"; exit 1;
+	fi
+
+	# get encryption key for decrypting
+	key=$(cut -d' ' -f1 < /etc/iris/swupdate/encryption.key)
+	iv=$(cut -d' ' -f2 < /etc/iris/swupdate/encryption.key)
+
+	# get current rootfs size
+	cur_size=$(lvs --select "lv_name = rootfs$FIRMWARE_SUFFIX" -o LV_SIZE --units B --nosuffix --noheadings | cut -c3-)
+	
+	# get new rootfs size
+	new_size=$(openssl enc -d -aes-256-cbc -K "$key" -iv "$iv" -in "$rootfs_file" | zcat | wc -c)
+	if [ $new_size -eq 0 ]; then
+		log "Could not retrieve new rootfs size during logical volume resize!"; exit 1;
+	fi
+
+	# mount over read-only /etc/lvm to modify config
+	if ! grep -qs /etc/lvm /proc/mounts; then
+		mkdir -p /tmp/etc/lvm
+		mount --bind /tmp/etc/lvm /etc/lvm
+	fi
+
+	log "Resize rootfs logical volume: ${cur_size} to ${new_size}"
+	lvresize --force --yes --quiet -L "$new_size"B "$ROOT_DEV" 2> /dev/null
+	vgmknodes
+
+	umount /etc/lvm
+	rm -R /tmp/etc
+}
+
 if [ "$1" = "preinst" ]; then
 	cmds_exist	
 	parse_cmdline
 	set_device_names
 	mount_keystore
+	resize_lvm
 	unlock_device
 	get_bootdev_name
 	create_symlinks